Patch from Gary (Bull) to remove warnings

+++ edu.rice.cs.hpc.data/src/edu/rice/cs/hpc/data/experiment/scope/AlienScope.java
@@ -99,7 +99,7 @@ public Scope duplicate() {
 			      this.procedureName,
 			      this.firstLineNumber, 
 			      this.lastLineNumber, 
-			      (int) this.cct_node_index);
+			      this.cct_node_index);
 
 }
